CryptoxHunter Q & A for 5 second bitcoin
Just a note to the skeptic dev in the Q&A : you said you are a software developer then you would understand a pointer (I would hope). We are trading an encrypted pointer to the BTC control functions for an encrypted address on chain. we just redirect the pointer to a new control address. The control address used to sign the control functions is our EGEM (Ethereum based) addressing. The EGEM signature triggers on chain functions that will sign the bitcoin address which is shown in the proof. This was the only way I could figure out how to post an address with encryption in plain view on chain. Its also not perfect yet (why testnet BTC) and we are releasing the code on our Egem Sapphire testnet as we go live. You are more than welcome to come and participate in the code development as that gets posted. Then if you think its a scam you can prove it in code and laugh at me all day long. Or, on the other hand if you think this is then ground breaking technology you can participate in it and list it on your resume as an open source implementation you helped work on for the entire crypto community. A truly decentralized multi asset DEX and Wallet with a super fast confirmation sounds like a worthy cause to me. I did not knock the lightning network in my answer to you because I originally based my approach on HTLC Alice and Bob thing and I think LN is cool AF. Just when doing it in conjunction with Sapphire I realized I already had POW nodes and I did not want to have an onramp time for each transaction. This is a tether to an ethereum based chain, based on its own blake2s POW, trading multiple assets through derived and network derived wallets from other chains. Where I thought LN would be my solution I ended up with what you see in my proof as a sensible solution to me, also a software developer and system architect with 15 years experience and a MS. I admit my solution may not be perfect - but who has a perfect solution right now anyway? (everyone gonna say me but truth is none of it is perfect we are all evolving). I am sharing this data with you and the community to evolve toward a more perfect solution. So if you are skeptical of the proof please, by all means, skeptic it in technical terms so we can use that feedback to improve the solution for future users of crypto currency. Any other software devs interested in working on something like this please contact me. It may take a few weeks (and my weeks can end up months) for me to get it posted out and get you working on it but that is my intention. Thanks.